refactor: Change timestamp format from seconds to milliseconds
Updated timestamp format in TTS output from floating-point seconds (e.g., 3.2)
to integer milliseconds (e.g., 3200) for better precision and consistency with
frontend requirements.

Changes:
- Modified TTSProcessor to output timestamps in milliseconds (integer)
- Updated PipelineOutput.total_duration type from float to int
- Updated API documentation in backend/README.md with new format examples
- Updated inference models documentation with millisecond format

Affected fields in timestamps.json:
- metadata.total_duration: 125.3 → 125300
- timestamps[].start_time: 0.0 → 0
- timestamps[].end_time: 3.2 → 3200
- timestamps[].duration: 3.2 → 3200

+++ backend/inference_models/lecture_pipeline.py
@@ -27,7 +27,7 @@ class PipelineOutput:
 
     # Metadata for client
     lecture_name: str = ""
-    total_duration: float = 0.0  # seconds
+    total_duration: int = 0  # milliseconds
     total_sentences: int = 0
     audio_sample_rate: int = 24000
